testing train_grid_and_save_best_model.py on kaggle env

+++ train_grid_and_save_best_model.py
@@ -41,7 +41,8 @@ def train_model(model_path, model_name):
 
     val_map = model.metrics.box.map
     print(f"map@50-95: {val_map}")
-    shutil.rmtree(RUN_DIR)
+    if os.path.exists(RUN_DIR):
+        shutil.rmtree(RUN_DIR)
     if val_map > current_best_map:
         shutil.rmtree(OUTPUT_DIR)
         print("Model updated")
